refactor: remove imp in serviceFilm
